nydusd: fix compilation errors related to unit test cases
Fix compilation errors related to unit test cases.
